==> notebooks/meetings/1_mp4_to_wav_batch_16khz.ipynb <==
# Batch Convert mp4 to wav

## Preparation

### Import

import os
from google.colab import drive
drive.mount('/content/drive')

## Function Definition

### Mp4 to wav batch

def mp4_to_wav_batch(input_dir,output_dir):
    files_processed=1
    for file in os.listdir(input_dir):  # iterate over files in input_dir
        if file.endswith('.mp4'):  # check if file has .mp4 extension
            input_path = os.path.join(input_dir, file)  # construct input path using os.path.join
            output_path = os.path.join(output_dir, os.path.splitext(file)[0] + '.wav')  # construct output path using os.path.join and os.path.splitext
            os.system(f'ffmpeg -i "{input_path}" -ac 1 -ar 16000 "{output_path}"')
            # run FFmpeg command to convert MP4 to WAV format with 1 audio channel and 16000 Hz sampling rate, and save output file at output_path
            print(str(files_processed)+") "+ str(output_path))
            files_processed = files_processed+1

## Use of Function

input_dir = r'/content/drive/MyDrive/Projects/tps/data/1. mp4'
output_dir = r'/content/drive/MyDrive/Projects/tps/data/2. wav'

mp4_to_wav_batch(input_dir,output_dir)
