==> notebooks/meetings/2_diarization_wav_to_rttm_batch.ipynb <==
# Diarization from wav to rttm in barch

## Preparation

### Install

!pip install -qq https://github.com/pyannote/pyannote-audio/archive/refs/heads/develop.zip
!pip install -qq ipython==7.34.0

### Import

from pyannote.audio import Pipeline
import os
from google.colab import drive
drive.mount('/content/drive')

### Download Model

huggingface_token='place_here_your_token'

pipeline = Pipeline.from_pretrained("pyannote/speaker-diarization@2.1",
                                    use_auth_token=huggingface_token)

## Function Definition

### Diarization

def diarization_rttm(input_file,output_file,num_speakers):
  diarization = pipeline(input_file,num_speakers=num_speakers)
  with open(output_file, "w") as rttm:
      diarization.write_rttm(rttm)

### Batch Diarization

def diarization_rttm_batch(directory_in,directory_out):
  os.chdir(directory_in)
  n=1
  for file in os.listdir():
      if file.endswith(".wav"):
          file_in = f"{directory_in}/{file}"
          file_out = os.path.join(directory_out, file[:-4] + ".rttm")
          if os.path.isfile(file_out) == False:
              num_speakers = dict_num_speakers[file]
              diarization_rttm(file_in,file_out,num_speakers)
              print(str(n)+") "+ str(file_out))
              n=n+1

## Use of Function

directory_in = r'/content/drive/MyDrive/Projects/tps/data/2. wav'
directory_out = r'/content/drive/MyDrive/Projects/tps/data/3. diarization (rttm)'
values= [5,4,5,5,6,6,6,4,4,4,5,5,5,4,4,4,4,4,6,6,4,4,4,4,5,6,5,6]
keys= ["1.1.wav","1.2.wav","2.1.wav","2.2.wav","3.1.wav","3.2.wav","3.3.wav","4.1.wav","4.2.wav","4.3.wav","5.1.wav","5.2.wav","6.1.wav","7.1.wav","7.2.wav","7.3.wav","7.4.wav","7.5.wav","8.1.wav","8.2.wav","9.1.wav","9.2.wav","10.1.wav","10.2.wav","11.1.wav","11.2.wav","12.1.wav","12.2.wav"]
dict_num_speakers = dict(zip(keys, values))

diarization_rttm_batch(directory_in,directory_out)
