==> notebooks/meetings/4_transcription_wav_to_srt_batch.ipynb <==
# Transcript Audio Files (wav to srt)

## Preparation

### Install

!pip install -U git+https://github.com/jianfch/stable-ts.git

### Import

import stable_whisper
import time
import os
from google.colab import drive
drive.mount('/content/drive')

### Download Model

model = stable_whisper.load_model('small')

## Function Definition

### Wav to Srt

def wav_to_srt(file_in,file_out):
    result = model.transcribe(file_in)
    result.to_srt_vtt(file_out, word_level=False)

### Wav to Srt Batch

def wav_to_srt_batch(directory_in,directory_out):
  # Moves to the directory and execute the function for every rttm file
  os.chdir(directory_in)
  n=0     
  for file in os.listdir():
      if file.endswith(".wav"):
          file_path = f"{directory_in}/{file}"
          file_out = os.path.join(directory_out, file[:-4] + ".srt")
          if os.path.isfile(file_out)==False:
              wav_to_srt(file_path,file_out)
              print(str(n)+") "+ str(file_out))
              n=n+1

## Use of Function

directory_in = r'/content/drive/MyDrive/Projects/tps/data/2. wav'
directory_out = r'/content/drive/MyDrive/Projects/tps/data/5. transcription (srt)/second_try'

wav_to_srt_batch(directory_in,directory_out)
