==> notebooks/meetings/10_emotion_text_bert_multilingual.ipynb <==
# Detect Emotions in Text using Bert Multilingual

## Preparation

### Install

!pip install torch
!pip install transformers

### Import

from transformers import BertTokenizer, BertForSequenceClassification, AdamW
import torch
from torch.utils.data import DataLoader, RandomSampler, SequentialSampler
from sklearn.metrics import f1_score
import pandas as pd
import os
from google.colab import drive
drive.mount('/content/drive')

### Download Model

tokenizer = BertTokenizer.from_pretrained('bert-base-multilingual-cased')

model = BertForSequenceClassification.from_pretrained('bert-base-multilingual-cased', num_labels=4)

# Define the emotion labels (assuming the labels are integers from 0 to 3, as in the previous example)
emotion_labels = ['joy', 'anger', 'fear', 'sadness']

## Function Definition

### Predict Emotions given a row in a Dataframe

def predict_all_emotion(row):

    text= row["Subtitle"]

    # Tokenize the input text and convert to PyTorch tensor
    input_ids = tokenizer.encode(text, add_special_tokens=True, padding='max_length', max_length=128, truncation=True, return_tensors='pt')
    
    # Predict the emotion label probabilities for the input text
    outputs = model(input_ids)
    label_probs = torch.softmax(outputs[0], dim=1).detach().numpy()[0]
    
    return {label: prob for label, prob in zip(emotion_labels, label_probs)}

### Analize Text Emotions given a csv

def analyze_emotions_csv(file_path, file_out):
  
  df = pd.read_csv(file_path, encoding="utf-8")
  df["emotions"] = df.apply(lambda row: predict_all_emotion(row), axis=1)

  # create new columns from emotions column
  df[['text_joy', 'text_anger', 'text_fear', 'text_sadness']] = df['emotions'].apply(lambda x: pd.Series(x))

  df = df.drop(columns=['emotions'])

  df.to_csv(file_out, index=False, encoding="utf-8-sig")

### Analize Text Emotions of csv in batch

def analyze_directory(directory_in, directory_out):
    # Create the output directory if it doesn't already exist
    os.makedirs(directory_out, exist_ok=True)

    # Iterate through all files in the input directory
    for file_name in os.listdir(directory_in):
        # Check if file is a CSV file
        if file_name.lower().endswith('.csv'):
            # Construct the input and output file paths
            file_path = os.path.join(directory_in, file_name)
            file_out = os.path.join(directory_out, os.path.splitext(file_name)[0] + '_text.csv')
            if os.path.exists(file_out)==False:
              # Analyze the CSV file and output the result
              analyze_emotions_csv(file_path, file_out)
              print(f"Processed {file_name}, saved output to {file_out}")

## Use of Function

directory_in = r'/content/drive/MyDrive/Projects/tps/data/10. speech_meetings_aggregated'
directory_out = r'/content/drive/MyDrive/Projects/tps/data/11. speech_text'

analyze_directory(directory_in, directory_out)
