==> notebooks/meetings/11_v2_renameIDs_unify_data.ipynb <==
# Individual Feature Determination

## Preparation

### Import

import pandas as pd
import os
import statistics
from google.colab import drive
drive.mount('/content/drive')

pd.set_option('display.max_rows', 100)

## Start

out_file = r'/content/drive/MyDrive/Projects/tps/meetings/data/11v2. speech_text_newID/speech_text_newID_unified.xlsx'
directory_in = r'/content/drive/MyDrive/Projects/tps/meetings/data/11. speech_text/'

def concatenate_csv_files(directory):
    # Get a list of all CSV files in the specified directory
    csv_files = [f for f in os.listdir(directory) if f.endswith('.csv')]

    # Initialize an empty list to store DataFrames
    dfs = []

    # Read and concatenate each CSV file
    for csv_file in csv_files:
        file_path = os.path.join(directory, csv_file)
        df = pd.read_csv(file_path)
        dfs.append(df)

    # Concatenate all DataFrames into one
    concatenated_df = pd.concat(dfs, ignore_index=True)

    return concatenated_df

### Test 1

df= concatenate_csv_files(directory_in)

df.head()

# Get unique values of the 'speaker' column and sort them
unique_speakers = sorted(df['Speaker'].unique())

# Print the sorted unique values
for speaker in unique_speakers:
    print(speaker)

len(unique_speakers)

### Test 2

directory_in = r'/content/drive/MyDrive/Projects/tps/meetings/data/8. dia_trans_verified_users/'
df= concatenate_csv_files(directory_in)

df.head()

# Get unique values of the 'speaker' column and sort them
unique_speakers = sorted(df['Speaker'].unique())

# Print the sorted unique values
for speaker in unique_speakers:
    print(speaker)

len(unique_speakers)
