==> notebooks/meetings/12_add_ratios_to_individual_feature.ipynb <==
# Individual Feature Determination

## Preparation

### Import

import pandas as pd
import numpy as np
from google.colab import drive
drive.mount('/content/drive')

## Function Definition

### Add Ratio given a column

def add_ratio_to_column(df, id_col, group_col, column, start_row):
  
    # Calculate the sum of spoken_time for each group
    group_sum = df.groupby(group_col)[column].transform('sum')

    # Calculate the spoken_time ratio
    ratio_column = column + "_ratio"
    column_index = df.columns.get_loc(column)

    # Insert the ratio_column next to the column received as a parameter
    df.insert(column_index + 1, ratio_column, np.nan)

    # Assign predefined values for the first three rows before start_row
    df.loc[start_row - 3:start_row - 1, ratio_column] = [df[column].iloc[start_row - 3],
                                                         "ratio of " + str(df[column].iloc[start_row - 2]),
                                                         "add_ratio_to_column()"]

    # Perform calculations for the ratio column from start_row onwards
    df.loc[start_row:, ratio_column] = df.loc[start_row:, column] / group_sum

    # Return the modified DataFrame
    return df



### Add Ratio given a Range

def add_ratios_range(df, start_col, end_col):

  cols_range = df.columns[df.columns.get_loc(start_col):df.columns.get_loc(end_col)+1].tolist()
    
  # Loop through the emotion columns and calculate the average for each
  for col in cols_range:
      df= add_ratio_to_column(df, "ID", "group", col, 3)

  return df

### Add Ratios to Features

def add_ratios_to_features(indiv_features_file, file_out):

  df= pd.read_excel(indiv_features_file)
  df= add_ratios_range(df, "indiv_spoken_time", "max_words_turn")
  df.to_csv(file_out,encoding='utf-8-sig', index=False)

  return df

## Use of Function

indiv_features_file = r'/content/drive/MyDrive/Projects/tps/data/12. features/3_individual_features_final.xlsx'
file_out= r'/content/drive/MyDrive/Projects/tps/data/12. features/4_individual_features_final_ratio.xlsx'
df= pd.read_excel(indiv_features_file)

df= add_ratios_to_features(indiv_features_file, file_out)
df.head()
