==> notebooks/meetings/13_group_features_meetings_determination.ipynb <==
# Individual Feature Determination

## Preparation

### Import

import pandas as pd
import numpy as np
from google.colab import drive
drive.mount('/content/drive')

pd.set_option('display.max_columns', 100)

## Function Definition

### Calculate Individual Features

#### Percentage Members Shown Face

def members_show_face_percen(df_group,df_indiv):

  grouped = df_indiv.groupby('group')
  dict_shown_face = {}

  for group, data in grouped:
      count = data['shown_face'].value_counts().get(True, 0)
      dict_shown_face[group] = count
  
  # Fill the DataFrame with the values from the dictionary
  df_group['members_show_face'] = df_group.index.map(dict_shown_face).fillna(0)
  df_group['members_show_face_percen'] = df_group['members_show_face'] / df_group['num_members']
  df_group= df_group.drop('members_show_face', axis=1)

  return df_group

#### Percentage Share Screen

def screen_share_percen(df_group):

  df_group['screen_share_percen'] = df_group['share_screen_time'] / df_group['total_duration_recorded']
  	
  return df_group


#### Overall Duration Spoken

def total_duration_spoken(df_group,df_indiv):

  grouped = df_indiv.groupby('group')
  spoken_duration = {}

  for group, data in grouped:
      sum = data['indiv_spoken_time'].sum()
      spoken_duration[group] = sum

  df_group['total_duration_spoken'] = df_group.index.map(spoken_duration).fillna(np.NaN)

  return df_group


#### Ratio Spoken-Recorded

def ratio_spoken_recorded(df_group,df_indiv):

  df_group['ratio_spoken_recorded'] = df_group['total_duration_spoken'] / df_group['total_duration_recorded']

  return df_group


#### Average Meeting Turn Duration

def avg_meeting_turn_duration(df_group, df_indiv):
  
    grouped = df_indiv.groupby('group')
    avg_durations = {}

    for group, data in grouped:
        turns_and_avg = []
        total = 0
        for index, row in data.iterrows():
            if pd.notna(row['num_turns']) and pd.notna(row['average_turn_duration']):
                turns_and_avg.append((row['num_turns'], row['average_turn_duration']))
                total += row['num_turns']
        if turns_and_avg:
            result = sum(pair[0] / total * pair[1] for pair in turns_and_avg)
            avg_durations[group] = result
    
    df_group['avg_meeting_turn_duration'] = df_group.index.map(avg_durations).fillna(np.NaN)

    return df_group

#### Total Turns Overall

def total_turns(df_group,df_indiv):

  grouped = df_indiv.groupby('group')
  turns = {}

  for group, data in grouped:
      sum = data['num_turns'].sum()
      turns[group] = sum

  df_group['total_turns'] = df_group.index.map(turns).fillna(np.NaN)

  return df_group


#### Turns per minute

def turns_in_a_min(df_group,df_indiv):

  df_group['turns_in_a_min'] = df_group['total_turns'] / df_group['total_duration_recorded']*60

  return df_group


#### Average meeting Words in Message

def average_meeting_words_message(df_group,df_indiv):

    grouped = df_indiv.groupby('group')
    avg_words = {}

    for group, data in grouped:
        words_and_avg = []
        total = 0
        for index, row in data.iterrows():
            if pd.notna(row['num_words']) and pd.notna(row['avg_words_turn']):
                words_and_avg.append((row['num_words'], row['avg_words_turn']))
                total += row['num_words']
        if words_and_avg:
            result = sum(pair[0] / total * pair[1] for pair in words_and_avg)
            avg_words[group] = result
    
    df_group['average_meeting_words_message'] = df_group.index.map(avg_words).fillna(np.NaN)

    return df_group

#### Total Words in General

def total_words(df_group,df_indiv):

  grouped = df_indiv.groupby('group')
  words = {}

  for group, data in grouped:
      sum = data['num_words'].sum()
      words[group] = sum

  df_group['total_words'] = df_group.index.map(words).fillna(np.NaN)

  return df_group


#### Words per minute

def words_in_a_min(df_group,df_indiv):

  df_group['words_in_a_min'] = df_group['total_words'] / df_group['total_duration_recorded']*60

  return df_group

### Calculate Descriptive Estadistics given a column

def calc_descriptive_stadistics_column2(df_group, df_indiv, group_col, column):

    grouped = df_indiv.groupby(group_col)
    turns = {}
    statistics_list= ['avg', 'stdev', 'min', 'max', 'diff']
    blank_dictionaries = {i: {} for i in statistics_list}

    for group, data in grouped:
      blank_dictionaries['avg'][group] = data[column].mean()
      blank_dictionaries['stdev'][group] = data[column].std()
      blank_dictionaries['min'][group] = data[column].min()
      blank_dictionaries['max'][group] = data[column].max()
      blank_dictionaries['diff'][group] = data[column].max() - data[column].min()

    for stats in statistics_list:
      df_group[stats+"_"+column] = blank_dictionaries[stats].values

    # Return the modified DataFrame
    return df_group

def calc_descriptive_statistics_column(df_group, df_indiv, group_col, column):
    grouped = df_indiv.groupby(group_col)
    statistics_list = ['avg', 'stdev', 'min', 'max', 'diff']
    blank_dictionaries = {i: {} for i in statistics_list}

    for group, data in grouped:
        blank_dictionaries['avg'][group] = data[column].mean()
        blank_dictionaries['stdev'][group] = data[column].std()
        blank_dictionaries['min'][group] = data[column].min()
        blank_dictionaries['max'][group] = data[column].max()
        blank_dictionaries['diff'][group] = data[column].max() - data[column].min()

    new_data = pd.DataFrame.from_dict(blank_dictionaries)

    # Rename the columns in new_data
    new_data.columns = [stats + "_" + column for stats in statistics_list]

    # Merge the new DataFrame with df_group
    df_group = pd.concat([df_group, new_data], axis=1)

    # Return the modified DataFrame
    return df_group


### Add Descriptive Estadistics given a Range

def add_descriptive_stadistics_range(df_group, df_indiv, start_col, end_col):

  cols_range = df_indiv.columns[df_indiv.columns.get_loc(start_col):df_indiv.columns.get_loc(end_col)+1].tolist()
    
  # Loop through the emotion columns and calculate the average for each
  for col in cols_range:
      df_group= calc_descriptive_statistics_column(df_group, df_indiv, "group", col)

  return df_group

### Calculate Group Features

def calc_group_features(indiv_features_file, group_features, file_out):

  df_indiv = pd.read_excel(indiv_features_file, index_col=0)
  df_group = pd.read_excel(group_features_file, index_col=0)
  df_group = df_group.transpose()

  stripped_rows = df_group.iloc[:4]  # Extract the first 4 rows and store them
  df_group = df_group.iloc[4:]
  df_indiv = df_indiv.iloc[3:]

  df_group= members_show_face_percen(df_group,df_indiv)
  df_group= screen_share_percen(df_group)
  df_group= total_duration_spoken(df_group,df_indiv)
  df_group= ratio_spoken_recorded(df_group,df_indiv)
  df_group= avg_meeting_turn_duration(df_group,df_indiv)
  df_group= total_turns(df_group,df_indiv)
  df_group= turns_in_a_min(df_group,df_indiv)
  df_group= average_meeting_words_message(df_group,df_indiv)
  df_group= total_words(df_group,df_indiv)
  df_group= words_in_a_min(df_group,df_indiv)
  
  df_group=add_descriptive_stadistics_range(df_group, df_indiv,"indiv_spoken_time", "text_sadness")
  
  #df_group = df_group.transpose()
  df_group = pd.concat([stripped_rows, df_group]).reset_index(drop=True)
  df_group.to_csv(file_out,encoding='utf-8-sig', index=False)

  return df_group

## Use of Function

indiv_features_file = r'/content/drive/MyDrive/Projects/tps/data/12. features/4_individual_features_final_ratio.xlsx'
group_features_file = r'/content/drive/MyDrive/Projects/tps/data/12. features/6_group_meetings_features_byhand_filled.xlsx'
file_out= r'/content/drive/MyDrive/Projects/tps/data/12. features/7_group_meetings_features_final.xlsx'

calc_group_features(indiv_features_file, group_features_file, file_out)
